kill server shell

+++ bak/kill_servers.sh
@@ -1,9 +1,11 @@
 #!/bin/sh 
 
 
-services="gateway  currency_service public_service token_service user_service wallet_service  kline_service "
-for service in $services;do 
+services="gateway  currency_service public_service token_service user_service wallet_service  kline_service  root"
+for service in $services;do
     service_id=`ps -ef |grep -v grep | grep $service | awk -F " "  '{print $2}'`
     echo "kill $service,$service_id"
-    kill -9 $service_id
+    if [ "$service_id" != "" ];then
+        kill -9 $service_id
+    fi
 done
